refact: Move ExpenseInfoSidebar to components
